requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==5.0.6
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
packaging==24.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-decouple==3.8
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.1
requests==2.32.3
service-identity==24.1.0
setuptools==70.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
zope.interface==6.4.post2
